🔨 Refactor: 컴포넌트명 변경 및 삼항연산자 렌더링 코드에서 제거
