+++ src/components/common/MapCurPosButton/index.ts
@@ -1,0 +1,1 @@
+export { default as MapCurPosButton } from './MapCurPosButton';
