--- src/components/common/MapPositionButton/index.ts
@@ -1,1 +1,0 @@
-export { default as MapPositionButton } from './MapPositionButton';
